+++ apps/explorer/src/index.tsx
@@ -15,8 +15,8 @@ import { growthbook } from './utils/growthbook';
 import { queryClient } from './utils/queryClient';
 import './utils/sentry';
 
-import './index.css';
 import '@mysten/dapp-kit/dist/index.css';
+import './index.css';
 
 // Load Amplitude as early as we can:
 initAmplitude();