+++ external-crates/move/crates/move-cli/src/base/test.rs
@@ -7,7 +7,7 @@ use anyhow::Result;
 use clap::*;
 use move_command_line_common::files::{FileHash, MOVE_COVERAGE_MAP_EXTENSION};
 use move_compiler::{
-    diagnostics::{self},
+    diagnostics::{self, Diagnostics},
     shared::{NumberFormat, NumericalAddress},
     unit_test::{plan_builder::construct_test_plan, TestPlan},
     PASS_CFGIR,
@@ -108,11 +108,10 @@ impl Test {
             cost_table,
             compute_coverage,
             &mut std::io::stdout(),
-            &mut std::io::stdout(),
         )?;
 
         // Return a non-zero exit code if any test failed
-        if let UnitTestResult::Failure = result {
+        if let (UnitTestResult::Failure, _) = result {
             std::process::exit(1)
         }
         Ok(())
@@ -126,16 +125,15 @@ pub enum UnitTestResult {
     Failure,
 }
 
-pub fn run_move_unit_tests<CW: Write + Send, TW: Write + Send>(
+pub fn run_move_unit_tests<W: Write + Send>(
     pkg_path: &Path,
     mut build_config: move_package::BuildConfig,
     mut unit_test_config: UnitTestingConfig,
     natives: Vec<NativeFunctionRecord>,
     cost_table: Option<CostTable>,
     compute_coverage: bool,
-    compiler_writer: &mut CW,
-    test_writer: &mut TW,
-) -> Result<UnitTestResult> {
+    writer: &mut W,
+) -> Result<(UnitTestResult, Option<Diagnostics>)> {
     let mut test_plan = None;
     build_config.test_mode = true;
     build_config.dev_mode = true;
@@ -179,7 +177,8 @@ pub fn run_move_unit_tests<CW: Write + Send, TW: Write + Send>(
     // Move package system, to first grab the compilation env, construct the test plan from it, and
     // then save it, before resuming the rest of the compilation and returning the results and
     // control back to the Move package system.
-    build_plan.compile_with_driver(compiler_writer, |compiler| {
+    let mut warning_diags = None;
+    build_plan.compile_with_driver(writer, |compiler| {
         let (files, comments_and_compiler_res) = compiler.run::<PASS_CFGIR>().unwrap();
         let (_, compiler) =
             diagnostics::unwrap_or_report_diagnostics(&files, comments_and_compiler_res);
@@ -190,8 +189,9 @@ pub fn run_move_unit_tests<CW: Write + Send, TW: Write + Send>(
         let compilation_result = compiler.at_cfgir(cfgir).build();
         let (units, warnings) =
             diagnostics::unwrap_or_report_diagnostics(&files, compilation_result);
-        diagnostics::report_warnings(&files, warnings);
+        diagnostics::report_warnings(&files, warnings.clone());
         test_plan = Some((built_test_plan, files.clone(), units.clone()));
+        warning_diags = Some(warnings);
         Ok((files, units))
     })?;
 
@@ -222,12 +222,12 @@ pub fn run_move_unit_tests<CW: Write + Send, TW: Write + Send>(
     // Run the tests. If any of the tests fail, then we don't produce a coverage report, so cleanup
     // the trace files.
     if !unit_test_config
-        .run_and_report_unit_tests(test_plan, Some(natives), cost_table, test_writer)
+        .run_and_report_unit_tests(test_plan, Some(natives), cost_table, writer)
         .unwrap()
         .1
     {
         cleanup_trace();
-        return Ok(UnitTestResult::Failure);
+        return Ok((UnitTestResult::Failure, warning_diags));
     }
 
     // Compute the coverage map. This will be used by other commands after this.
@@ -235,7 +235,7 @@ pub fn run_move_unit_tests<CW: Write + Send, TW: Write + Send>(
         let coverage_map = CoverageMap::from_trace_file(trace_path);
         output_map_to_file(coverage_map_path, &coverage_map).unwrap();
     }
-    Ok(UnitTestResult::Success)
+    Ok((UnitTestResult::Success, warning_diags))
 }
 
 impl From<UnitTestResult> for ExitStatus {