+++ external-crates/move/crates/move-stdlib/tests/move_unit_test.rs
@@ -38,10 +38,9 @@ fn run_tests_for_pkg(path_to_pkg: impl Into<String>, include_nursery_natives: bo
         None,
         /* compute_coverage */ false,
         &mut std::io::stdout(),
-        &mut std::io::stdout(),
     )
     .unwrap();
-    if result != UnitTestResult::Success {
+    if result.0 != UnitTestResult::Success {
         panic!("aborting because of Move unit test failures");
     }
 }