+++ sdk/dapp-kit/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mysten/dapp-kit
 
+## 0.10.2
+
+### Patch Changes
+
+- 2ae1a6a108: Early return on `useAutoConnectWallet` to improve account switching performance
+- b30fecbbca: Fix missing export of useSuiClientQueries
+- Updated dependencies [cdcfa76c43]
+  - @mysten/sui.js@0.48.0
+  - @mysten/wallet-standard@0.8.11
+
 ## 0.10.1
 
 ### Patch Changes